Remove calls from interpret, remove lookup.rs
Remove calls from interpret since it is just focusing on definitions,
removed lookup.rs in favor of putting it into interpret.rs.

Renamed NASLInterpreter and OpenVASInterpreter to NASLDefinitions and
OpenVASBuildInFunctions.
