+++ nasl/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod cache;
 pub mod interpret;
-mod lookup;
 pub mod openvas_funcs;
 mod node_ext;
 pub mod types;
